+++ client_code/zod/__init__.py
@@ -1,7 +1,21 @@
 # SPDX-License-Identifier: MIT
 # Copyright (c) 2021 anvilistas
 
-from ._types import any_ as any
-from ._types import boolean, literal, never, none, string, unknown
+from ._types import (
+    any,
+    boolean,
+    date,
+    datetime,
+    float,
+    integer,
+    literal,
+    never,
+    none,
+    number,
+    object,
+    string,
+    union,
+    unknown,
+)
 
 __version__ = "0.0.1"
